backend/requirements.txt
# List of dependencies
fastapi[all]
pandas
openai
python-dotenv
todoist_api_python